# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CHASSII API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts with filters
      parameters:
        - name: make
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, popular, unanswered, local, following]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post with comments
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostDetail"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{postId}/save:
    post:
      operationId: savePost
      tags: [posts]
      summary: Save/bookmark a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
    delete:
      operationId: unsavePost
      tags: [posts]
      summary: Unsave a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unsaved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResult"
  /posts/{postId}/comments:
    get:
      operationId: getPostComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ── Feed ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, displayName, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        isFollowingMe:
          type: boolean
        iFollowThem:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        carCount:
          type: integer
        postCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Post:
      type: object
      required: [id, userId, title, category, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        carId:
          type: ["integer", "null"]
        title:
          type: string
        body:
          type: string
        category:
          type: string
          enum: [general, question, build_update, guide, review, problem, event, marketplace]
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        generation:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        author:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    PostDetail:
      type: object
      required: [id, userId, title, category, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        carId:
          type: ["integer", "null"]
        title:
          type: string
        body:
          type: string
        category:
          type: string
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        generation:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        commentCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        author:
          $ref: "#/components/schemas/User"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [title, body, category]
      properties:
        title:
          type: string
        body:
          type: string
        category:
          type: string
          enum: [general, question, build_update, guide, review, problem, event, marketplace]
        carId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        generation:
          type: string
        location:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrls:
          type: array
          items:
            type: string
    Comment:
      type: object
      required: [id, postId, userId, body, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        parentId:
          type: ["integer", "null"]
        body:
          type: string
        likeCount:
          type: integer
        isLiked:
          type: boolean
        author:
          $ref: "#/components/schemas/User"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
        parentId:
          type: integer
...
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
    SaveResult:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
...