# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CHASSII API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface Comment {
  id: number;
  postId: number;
  userId: number;
  /** @nullable */
  parentId?: number | null;
  body: string;
  likeCount?: number;
  isLiked?: boolean;
  author?: User;
  replies?: Comment[];
  createdAt: Date;
}