# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CHASSII API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  title?: string;
  body?: string;
  tags?: string[];
  imageUrls?: string[];
}